Refactor StudentCard into shared component
